airflow: move version import
