+++ integration/airflow/openlineage/airflow/facets.py
@@ -7,7 +7,7 @@ import attr
 from airflow.version import version as AIRFLOW_VERSION
 from openlineage.client.facet import BaseFacet
 
-from openlineage.airflow import __version__ as OPENLINEAGE_AIRFLOW_VERSION
+from openlineage.airflow.version import __version__ as OPENLINEAGE_AIRFLOW_VERSION
 from openlineage.client.utils import RedactMixin
