+++ integration/airflow/openlineage/airflow/version.py
@@ -1,0 +1,1 @@
+__version__ = "0.13.0"
