+++ integration/airflow/setup.cfg
@@ -11,7 +11,7 @@ serialize =
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
 
-[bumpversion:file:openlineage/airflow/__init__.py]
+[bumpversion:file:openlineage/airflow/version.py]
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
 
@@ -23,6 +23,7 @@ addopts = -p no:warnings
 
 [mypy]
 ignore_missing_imports = True
+disable_error_code = attr-defined
 
 [tox:tox]
 envlist =
