+++ integration/airflow/tests/test_openlineage_dag.py
@@ -29,7 +29,7 @@ from openlineage.common.models import (
 )
 from openlineage.common.sql import DbTableMeta
 from openlineage.airflow import DAG
-from openlineage.airflow import __version__ as OPENLINEAGE_AIRFLOW_VERSION
+from openlineage.airflow.version import __version__ as OPENLINEAGE_AIRFLOW_VERSION
 from openlineage.airflow.extractors import (
     BaseExtractor, TaskMetadata
 )
